requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
idna==3.2
mutagen==1.45.1
pafy==0.5.5
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
spotipy==2.19.0
urllib3==1.26.7
youtube-dl==2021.6.6
